Теперь в WeatherModel можно получить название города и страну. Значения температуры, ветра и давления теперь double

+++ app/src/main/java/ru/geekbrains/androidBase/lesson1/db/DataSource.java
@@ -39,7 +39,7 @@ public class DataSource implements Closeable {
         AppSettingsSingleton.getInstance().setDataSource(null);
     }
 
-    public void updateHistory(String city, int temperature, int wind, int pressure){
+    public void updateHistory(String city, double temperature, double wind, double pressure){
         //move city name to lower case
         String cityString = city.toLowerCase();
         //get nows date
